More consistent use of size_t (instead of int) for counting and indexing things.
These changes were prompted by deprecation warnings issued by the new
DMD v2.084.0 compiler.  It is getting more picky -- a good thing.

Quite a few small changes were made to Brad's electronically-specific gas model
and these need to be reviewed.

+++ src/gas/gas_model.d
@@ -63,7 +63,7 @@ public:
     @nogc @property ref double[] LJ_epsilons() { return _LJ_epsilons; }
     @nogc @property ref double[] Le() { return _Lewis_numbers; }
     @nogc @property ref double[] electronic_energy() { return _electronic_energy; }
-    @nogc final string species_name(int i) const { return _species_names[i]; }
+    @nogc final string species_name(size_t i) const { return _species_names[i]; }
     final int species_index(string spName) const { return _species_indices.get(spName, -1); }
     @nogc final string energy_mode_name(int i) const { return _energy_mode_names[i]; }
     final int energy_mode_index(string modeName) const {
